renamed user settings page to account page;
