+++ lib/presentation/pages/account_page/widgets/account_list_item.dart
@@ -1,16 +1,16 @@
 import 'package:blogify_flutter_main/common/app_colors.dart';
 import 'package:blogify_flutter_main/common/app_dimensions.dart';
-import 'package:blogify_flutter_main/presentation/pages/user_settings_page/widgets/section_title.dart';
+import 'package:blogify_flutter_main/presentation/pages/account_page/widgets/section_title.dart';
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 
-class ListItem extends StatelessWidget {
+class AccountListItem extends StatelessWidget {
   final IconData icon;
   final String title;
   final Function()? onTap;
   final double borderRadius;
 
-  const ListItem({
+  const AccountListItem({
     required this.title,
     required this.icon,
     this.onTap,
@@ -28,15 +28,13 @@ class ListItem extends StatelessWidget {
         onTap: onTap,
         child: Padding(
           padding: const EdgeInsets.symmetric(
-              vertical: AppDimensions.normalS, horizontal: AppDimensions.normalM),
+            vertical: AppDimensions.normalS,
+            horizontal: AppDimensions.normalM,
+          ),
           child: Row(
             spacing: AppDimensions.minorL,
             children: [
-              FaIcon(
-                icon,
-                size: AppDimensions.appBarIconSize,
-                color: Colors.black87,
-              ),
+              FaIcon(icon, size: AppDimensions.appBarIconSize, color: Colors.black87),
               SectionTitle(title: title),
             ],
           ),
