+++ lib/router/router.gr.dart
@@ -10,11 +10,14 @@
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
 import 'package:auto_route/auto_route.dart' as _i5;
-import 'package:blogify_flutter_main/presentation/pages/article_page/article_page.dart' as _i1;
-import 'package:blogify_flutter_main/presentation/pages/comments_page/comments_page.dart' as _i2;
-import 'package:blogify_flutter_main/presentation/pages/home_page/home_page.dart' as _i3;
-import 'package:blogify_flutter_main/presentation/pages/user_settings_page/user_settings_page.dart'
+import 'package:blogify_flutter_main/presentation/pages/account_page/account_page.dart'
     as _i4;
+import 'package:blogify_flutter_main/presentation/pages/article_page/article_page.dart'
+    as _i1;
+import 'package:blogify_flutter_main/presentation/pages/comments_page/comments_page.dart'
+    as _i2;
+import 'package:blogify_flutter_main/presentation/pages/home_page/home_page.dart'
+    as _i3;
 import 'package:flutter/material.dart' as _i6;
 
 /// generated route for
@@ -25,10 +28,10 @@ class ArticleRoute extends _i5.PageRouteInfo<ArticleRouteArgs> {
     _i6.Key? key,
     List<_i5.PageRouteInfo>? children,
   }) : super(
-          ArticleRoute.name,
-          args: ArticleRouteArgs(id: id, key: key),
-          initialChildren: children,
-        );
+         ArticleRoute.name,
+         args: ArticleRouteArgs(id: id, key: key),
+         initialChildren: children,
+       );
 
   static const String name = 'ArticleRoute';
 
@@ -62,10 +65,10 @@ class CommentsRoute extends _i5.PageRouteInfo<CommentsRouteArgs> {
     _i6.Key? key,
     List<_i5.PageRouteInfo>? children,
   }) : super(
-          CommentsRoute.name,
-          args: CommentsRouteArgs(id: id, key: key),
-          initialChildren: children,
-        );
+         CommentsRoute.name,
+         args: CommentsRouteArgs(id: id, key: key),
+         initialChildren: children,
+       );
 
   static const String name = 'CommentsRoute';
 
@@ -95,7 +98,7 @@ class CommentsRouteArgs {
 /// [_i3.HomePage]
 class HomeRoute extends _i5.PageRouteInfo<void> {
   const HomeRoute({List<_i5.PageRouteInfo>? children})
-      : super(HomeRoute.name, initialChildren: children);
+    : super(HomeRoute.name, initialChildren: children);
 
   static const String name = 'HomeRoute';
 
@@ -111,7 +114,7 @@ class HomeRoute extends _i5.PageRouteInfo<void> {
 /// [_i4.UserSettingsPage]
 class UserSettingsRoute extends _i5.PageRouteInfo<void> {
   const UserSettingsRoute({List<_i5.PageRouteInfo>? children})
-      : super(UserSettingsRoute.name, initialChildren: children);
+    : super(UserSettingsRoute.name, initialChildren: children);
 
   static const String name = 'UserSettingsRoute';
